added a few more verifications for Item

+++ backend/src/main/java/com/example/backend/item/ItemService.java
@@ -1,5 +1,7 @@
 package com.example.backend.item;
 
+import com.example.backend.exception.ErrorMessage;
+import com.example.backend.exception.SchrankException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -28,16 +30,20 @@ public class ItemService {
     public ItemDTO getItemById(String id) {
         return repository.findById(id)
                 .map(mapper::toDTO)
-                .orElse(null);
+                .orElseThrow(() -> new SchrankException(ErrorMessage.ITEM_NOT_FOUND, id));
     }
 
     public ItemDTO createItem(ItemDTO itemDTO) {
         ItemEntity itemEntity = mapper.toEntity(itemDTO);
+        itemEntity.verifyInvariants();
         ItemEntity savedEntity = repository.save(itemEntity);
         return mapper.toDTO(savedEntity);
     }
 
     public void deleteItem(String id) {
+        if (!repository.existsById(id)) {
+            throw new SchrankException(ErrorMessage.ITEM_NOT_FOUND, id);
+        }
         repository.deleteById(id);
     }
 
@@ -54,6 +60,6 @@ public class ItemService {
                     ItemEntity updatedEntity = repository.save(existingEntity);
                     return mapper.toDTO(updatedEntity);
                 })
-                .orElse(null);
+                .orElseThrow(() -> new SchrankException(ErrorMessage.ITEM_NOT_FOUND, id));
     }
 }
